VendingMachineUserMode,ManagerMode 객체 생성
